Restructure index page html.
The main reason for this is that it causes long titles to have less
line spacing when they wrap, which makes it clear that they're part of
the same title, rather than two different posts.

A secondary reason is that it maybe makes the html more "semantic" or
whatever bullshit.